requirements.txt
# OpenLLM Toolkit - Python Dependencies

# Core dependencies
asyncio
aiohttp>=3.8.0
requests>=2.28.0
python-dotenv>=0.19.0
colorama>=0.4.4

# MCP (Model Context Protocol)
mcp>=1.0.0

# Image processing
Pillow>=9.0.0
opencv-python>=4.6.0
pytesseract>=0.3.10

# HTTP and networking
httpx>=0.24.0
websockets>=10.0

# Data processing
pandas>=1.5.0
numpy>=1.21.0

# Utilities
rich>=12.0.0
click>=8.0.0
typer>=0.7.0
pydantic>=1.10.0

# Optional AI/ML libraries
torch>=1.12.0
transformers>=4.20.0
accelerate>=0.20.0

# Web backend
fastapi>=0.100.0
uvicorn>=0.20.0

# Development and testing
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0